refactor: split peer routing and send outcomes
